server/requirements.txt
aiofiles==0.6.0
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
click==7.1.2
fastapi==0.63.0
h11==0.12.0
httptools==0.1.1
idna==3.1
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
Pillow==8.1.0
pydantic==1.7.3
python-dotenv==0.15.0
PyYAML==5.4.1
SQLAlchemy==1.3.23
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.3
uvloop==0.14.0
watchgod==0.6
websockets==8.1
yarl==1.6.3
